Add timezone to elasticsearch range date filter
This commit adds a time_zone argument to the Elasticsearch query, as
spree_events.start_time is stored with a time_zone rather than UTC
format.

If Leo Cussen decide to offer courses outside of Victoria, this will 
allow the search form to include a form field with the timezone for 
the selected state, rather than hard-coding ‘Melbourne/Australia’.

We also change the filter keys :gte and :lte to :to and :from which
Elasticsearch supports for specifying a date range.